+++ process/process_posix.go
@@ -114,12 +114,6 @@ func PidExistsWithContext(ctx context.Context, pid int32) (bool, error) {
 	return false, err
 }
 
-// SendSignal sends a unix.Signal to the process.
-// Currently, SIGSTOP, SIGCONT, SIGTERM and SIGKILL are supported.
-func (p *Process) SendSignal(sig syscall.Signal) error {
-	return p.SendSignalWithContext(context.Background(), sig)
-}
-
 func (p *Process) SendSignalWithContext(ctx context.Context, sig syscall.Signal) error {
 	process, err := os.FindProcess(int(p.Pid))
 	if err != nil {
@@ -134,49 +128,24 @@ func (p *Process) SendSignalWithContext(ctx context.Context, sig syscall.Signal)
 	return nil
 }
 
-// Suspend sends SIGSTOP to the process.
-func (p *Process) Suspend() error {
-	return p.SuspendWithContext(context.Background())
-}
-
 func (p *Process) SuspendWithContext(ctx context.Context) error {
-	return p.SendSignal(unix.SIGSTOP)
-}
-
-// Resume sends SIGCONT to the process.
-func (p *Process) Resume() error {
-	return p.ResumeWithContext(context.Background())
+	return p.SendSignalWithContext(ctx, unix.SIGSTOP)
 }
 
 func (p *Process) ResumeWithContext(ctx context.Context) error {
-	return p.SendSignal(unix.SIGCONT)
-}
-
-// Terminate sends SIGTERM to the process.
-func (p *Process) Terminate() error {
-	return p.TerminateWithContext(context.Background())
+	return p.SendSignalWithContext(ctx, unix.SIGCONT)
 }
 
 func (p *Process) TerminateWithContext(ctx context.Context) error {
-	return p.SendSignal(unix.SIGTERM)
-}
-
-// Kill sends SIGKILL to the process.
-func (p *Process) Kill() error {
-	return p.KillWithContext(context.Background())
+	return p.SendSignalWithContext(ctx, unix.SIGTERM)
 }
 
 func (p *Process) KillWithContext(ctx context.Context) error {
-	return p.SendSignal(unix.SIGKILL)
-}
-
-// Username returns a username of the process.
-func (p *Process) Username() (string, error) {
-	return p.UsernameWithContext(context.Background())
+	return p.SendSignalWithContext(ctx, unix.SIGKILL)
 }
 
 func (p *Process) UsernameWithContext(ctx context.Context) (string, error) {
-	uids, err := p.Uids()
+	uids, err := p.UidsWithContext(ctx)
 	if err != nil {
 		return "", err
 	}
